requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
requests==2.31.0
loguru==0.7.2
python-dotenv==1.0.1
langchain==0.1.9
langchain-groq==0.0.1
regex==2023.12.25
httpx==0.27.0